Use smarter logic for avalara_tax_enabled on order
Doing a global check isn't great, as we can't run two tax system's side
by side without them stepping over each other.

This works to move the checks on avalara to "local logic" on the order,
which should only use avalara if it has a transaction or has any taxes
that are avalara calculated.