Fix potential error if cache_key is nil
If the order doesn't have a ship_address set, `order.tax_address` will
return a `Spree::Tax::TaxLocation` that doesn't respond to `cache_key`.
If we try to append nil to a string we'll get a TypeError. Calling to_s
ensures if this is nil for some reason that we just append an empty
string.