Add tests for requesting tax with discounts
There is a bug currently where we essentially double send shipment
discounts. It gets included in the invoice discount (which gets
distributed amongst `discounted?` items, but then we also pass the
discounted shipment item for the individual item.

The skipped test here highlights the problematic case and will be fixed
in a subsequent commit.