Don't error on non-calculated adjustments
The safe navigation operator `&.` checks if the object responds to the
method and throws an error if it doesn't. Since not all adjustment
sources include the Spree::CalculatedAdjustments module, they don't all
respond to `.calculator`. One such example is the free shipping
promotion action. These adjustments would cause this method to throw an
error and blow up order completion because we check this on the
payment's transition to complete.