Send tax date override for completed orders
This is already done for the return_tax request, but not for the
standard get_tax request.

Since Solidus recalculates taxes on every order update, if the rates
change and the order is updated after it completes (ie: when it ships),
the taxes will be recalculated at the new rate and the order will either
end up in `balance_due` or `credit_owed`.

This change ensures that we always send a tax date override for orders
that have a `completed_at` timestamp set.