Don't compute tax using expired tax rates
This calculator inherits from `Spree::Calculator::DefaultTax`, which
always returns 0 when trying to compute an item or order using an
expired tax rate.

However, this gem overrides the `compute_` methods with
`compute_shipment_or_line_item`, which removes that behaviour. I don't
believe there is a valid reason to change this, so I'm adding
it to the override here.